+++ src/partials/home/exercises.html
@@ -63,16 +63,6 @@
       </div>
     </div>
 
-    <div class="exercises-list-wrapper">
-      <div class="exercises-list-container"></div>
-
-      <div class="pagination">
-        <button class="pagination-button">1</button>
-        <button class="pagination-button active">2</button>
-        <button class="pagination-button">3</button>
-      </div>
-    </div>
-
     <div class="exercises-quote-container">
       <load ="partials/common/quote.html" />
     </div>
